Update record view for GDT
Why these changes are being introduced:

The GeoData app has specific design requirements for the record
view.

Relevant ticket(s):

* [GDT-131](https://mitlibraries.atlassian.net/browse/GDT-131)
* [GDT-130](https://mitlibraries.atlassian.net/browse/GDT-130)

How this addresses that need:

This adds a `record_geo` partial that includes metadata specific to
GDT records, such as locations and a metadata download link. It
also changes where the `access_button` partial is displayed
depending on screen width, and it adds a 'back' button to the top
of the record if `url_for(:back)` resembles a search results page.

Side effects of this change:

* Some of the changes introduced in GDT-130  have been overwritten
by this commit, as certain styles and markup are shared by both views.
Specifically, the `geo_data_info` and `_authors` partials have been
moved to the shared directory, and a new `shared` SCSS partial
contains styles used in `result` and `record`.
* Several helper methods have been added to parse the metadata needed
for the record view.