Address UX feedback on location-based search form
Why these changes are being introduced:

This is in response to a UX review of location-based search.

Relevant ticket(s):

* [GDT-136](https://mitlibraries.atlassian.net/browse/GDT-136)

How this addresses that need:

* Removes top margin from all `details` elements except the first.
* Bolds field labels and normalizes font size for all labels (including
hints and descriptions).
* Moves placeholder text to description `span` elements and changes
"e.g." to "Ex:".
* Changes language for distance description/hint.
* Changes summary labels for geobox and geodistance.
* Changes red asterisk 'required field' indicators to black as indicators,
and adds a `p` tag explaining what the asterisks mean beneat each `legend`.

Side effects of this change:

None.