Enable access to files filter
Why these changes are being introduced:

GDT requires an access to files filter, which allows users to
facet based on access to geodata files.

Relevant ticket(s):

* [GDT-140](https://mitlibraries.atlassian.net/browse/GDT-140)

How this addresses that need:

This enables the filter in the same manner we've established for
other filters:

* Adds it to the the FILTER_PARAMS constants in the Enhancer and
QueryBuilder models
* Adds it to the `nice_labels` method
* Includes it in the TIMDEX search heredoc

Side effects of this change:

* There is an additional requirement in GDT-140 to render the filter
as checkboxes. That introduces a fair amount of complexity into
the view layer, so it made sense to make that change separate for ease
of code review.
* We likely don't want this filter to appear in non-GDT apps, so we
should update the `ACTIVE_FILTERS` env in those apps accordingly
before this lands.
* The instructions for the `ACTIVE_FILTERS` env in the README have
been updated for clarity.
* Cassettes have been regenerated to account for the new filter.