Iterate on PR template
Why these changes are being introduced:

EngX is considering revamping our PR templates to make them more
useful. TIMDEX UI is a good starting point to try this out, as
it's under active development and sees many PRs.

Relevant ticket(s):

* [ENGX-226](https://mitlibraries.atlassian.net/browse/ENGX-226)

How this addresses that need:

This is a first effort to make the TIMDEX UI template more useful.
The checklist remains fairly general, but seeks to resolve some of
the issues of ambiguity that we've discussed. E.g., checking "The
documentation has been updated or is unnecessary" could mean two
different things.

Side effects of this change:

This will add some friction to the code review process, both on
the side of the PR author and the reviewer. We'll want to evaluate
after a few PRs whether this is too little or too much detail.