Remove border from result hover/focus
Why these changes are being introduced:

The border causes the result container to shift 1px to the right
on hover/focus.

Relevant ticket(s):

* [GDT-237](https://mitlibraries.atlassian.net/browse/GDT-237)
* [GDT-215](https://mitlibraries.atlassian.net/browse/GDT-215)

How this addresses that need:

This removes the border, but retains the background color change
so there is still an indication of focus.

Side effects of this change:

None.