Update theme gem and add skip_links partial
Why these changes are being introduced:

We are trying out custom skip links for this application, as
it has a relatively complex navigation structure. The theme gem
has been updated to support this, using a new `_skip_links` partial.

Relevant ticket(s):

* [GDT-161](https://mitlibraries.atlassian.net/browse/GDT-161)

How this addresses that need:

This updates the theme gem and adds the new partial as-is. It will
be customized in a subsequent commit.

Side effects of this change:

A small amount of technical debt may be added here, but incorporating
a layout that needs to remain compatible with the upstream theme
gem. However, the partial itself is very simple, so it's unlikely
that any maintenance will be required.