Make filter sidebar more semantically meaningful
Why these changes are being introduced:

VoiceOver (and possibly other screen readers) do not interpret
the current sidebar drawers as disclosure triangles. As such,
one must tab through every available filter, even if that
filter category is collapsed.

Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/GDT-161

How this addresses that need:

This refactors filter categories as `details` tags, so screen readers
will recognize them as expandable elements.

Side effects of this change:

* Some tests have been updated a result of changes to class names.
* We had previously been opening/closing categories with an `onclick`,
was a quick-and-dirty way to get things working but not very good
practice. I took this opportunity to change that too an event
listener and moved it to the `filters.js` file that we added for
the mobile filter toggle button.
* I've removed the transition rules for the filters, since `details`
does not support them natively.