Adjust font size for authors element
Why these changes are being introduced:

This is a follow-on request for GDT-238 that Darcy suggested and
I enthusiastically support.

Relevant ticket(s):

* [GDT-238](https://mitlibraries.atlassian.net/browse/GDT-238)

How this addresses that need:

This reduces the font size for authors to 1.6rem.

Side effects of this change:

While reviewing this, Darcy noticed a bug in iOS where the second
line of an ordered list element appears to have top padding. We
opened [GDT-246](https://mitlibraries.atlassian.net/browse/GDT-246)
to explore this further.