Change full record skip link text
Why these changes are being introduced:

UXWS had originally requested 'skip to metadata' for this link,
but decided on review that 'skip to content' would be preferable.

Relevant ticket(s):

* [GDT-161](https://mitlibraries.atlassian.net/browse/GDT-161)

How this addresses that need:

This changes the 'skip to metadata' link text to 'skip to content'.

Side effects of this change:

None.