Relabel source values for GDT
Why these changes are being introduced:

GDT needs to provide users the option to toggle between MIT
and non-MIT results. The source filter is reasonable solution
for MVP, but the current values are not intuitively named.

Relevant ticket(s):

* [GDT-245](https://mitlibraries.atlassian.net/browse/GDT-245)

How this addresses that need:

This adds a `gdt_sources` method in the filter helper that
that translates the two GDT sources, or returns the value
as-is otherwise. If the GDT feature is enabled, then source
values are translated to the desired labels in the filter
panel and sidebar.

Side effects of this change:

There was a lingering `to_sym` coercion in the filter partial.
I removed it, as it is no longer need after last month's filter
redesign. Its removal is unrelated to this change.