Address UX feedback on GDT result view
Why these changes are being introduced:

This responds to UX feedback on GDT-130 (result view for GDT).

Relevant ticket(s):

* [GDT-130](https://mitlibraries.atlassian.net/browse/GDT-130)

How this addresses that need:

This removes result padding in smaller viewports, moves result markers
outside of the titles, changes full record link hover/focus color to blue,
and removes the visited styling for full record links.

The visited styling remains for 'View online' links in other TIMDEX UI
apps, as these are outlinks to the source record. The GDT links go to
the full record, which is already indicated as visited in the title link,
so it doesn't feel critical to add visited styling to the button as well.

Side effects of this change:

[GDT-248](https://mitlibraries.atlassian.net/browse/GDT-248) has been
opened to address some feedback on the result highlight labels.