Add more fields to the GDT full record view
Why these changes are being introduced:

The 'more information' section of a GDT record should contain
all available fields, not just a selection. This is required
for initial usability testing, so stakeholders can decide
what fields they want to include.

Relevant ticket(s):

* [GDT-131](https://mitlibraries.atlassian.net/browse/GDT-131)

How this addresses that need:

This adds all the fields that I suspect will appear in a GIS
record. It intentionally omits some, like holdings and
callNumbers, that I suspect will not. The display is intentionally
messy at this point, as not all of these fields will be in the
final product.

Side effects of this change:

I left in some convenience methods (and corresponding tests) that
I suspect will be useful at some point, even though they will not
be used in the current testing phase. I can remove these if their
presence is more confusing than helpful.