Move results context from filter sidebar to results container
Why these changes are being introduced:

We will be moving the filter sidebar further down the tab order,
so screen reader users would not reach the number of search results
until they reach the end of the results. This foregrounds that
context, both for blind/low-vision and sighted users.

Relevant ticket(s):

* [GDT-235](https://mitlibraries.atlassian.net/browse/GDT-235)

How this addresses that need:

This moves the number of results returned to the top of the
results container, making it the first focusable element when
a screen reader user skips to results.

Side effects of this change:

The results ID moves up one level in the DOM to account for the
new heading.