Move GDT search controller tests to a separate file
Why these changes are being introduced:

Having these tests inherit from SearchControllerTest was causing
erratic behavior with regard to the test environment (i.e., the
`:gdt` feature was toggled on when it was not supposed to be).

Relevant ticket(s):

Discovered while testing [GDT-219](https://mitlibraries.atlassian.net/browse/GDT-219)

How this addresses that need:

This isolates the GDT search controller tests in a separate file
and explicitly disables `:gdt` at the top of SearchControllerTest.

Side effects of this change:

None.