Reorganize advanced search form for GDT
Why these changes are being introduced:

Stakeholders have requested that 'citation', 'funding information',
and 'identifiers' be removed from the advanced search form, and that
'contributors' be renamed to 'authors'. UXWS has also requested that
we hide the 'source' field in GDT, as we now only have two sources
and have effectively turned that filter into a toggle.

Relevant ticket(s):

* [GDT-219](https://mitlibraries.atlassian.net/browse/GDT-219)

How this addresses that need:

This renders the unwanted fields only if the `:gdt` feature is
not enabled, and similarly changes the label of 'contributors' to
'authors' when `:gdt` is enabled.

It also moves 'title' to the top of the form (and citation lower)
for both GDT and non-GDT contexts, as this seems like a useful
change for any TIMDEX UI application.

Side effects of this change:

Some of the markup has been reformatted to match our line width
limits.