Update access type values
Why these changes are being introduced:

The "access to files" values have been changed in the data mapping
layer. We need to make a corresponding change in a few partials
where the value of that field is a rendering condition.

Relevant ticket(s):

* [GDT-252](https://mitlibraries.atlassian.net/browse/GDT-252)

How this addresses that need:

This updates the values as needed in the access_button, record_geo, and
geo_data_info partials, as well as the access_type helper method.

Side effects of this change:

It was a bit tricky to wrangle and update all of these hardcoded values.
Though they are unlikely to change again, this is something that
we may want to consider refactoring.