Fix bug in filter sidebar SCSS
Why these changes are being introduced:

The SCSS for the filter names in the sidebar doesn't work exactly
the way we intend it to. When an applied filter is in focus, the
font color is black; it should be white to contrast with the
background.

Relevant ticket(s):

Discovered while working on [GDT-232](https://mitlibraries.atlassian.net/browse/GDT-232),
but otherwise unrelated to that ticket.

How this addresses that need:

This updates the SCSS so that applied filters are white on focus
and hover.

Side effects of this change:

None.