Remove skips from record controller tests
Why these changes are being introduced:

There were three record controller tests that had to be skipped due
to a data reload that was in process. That reload, which was
related to the access type values, is now complete, and the
corresponding UI code has been updated, so these tests will pass again.

Relevant ticket(s):

* [GDT-252](https://mitlibraries.atlassian.net/browse/GDT-252)

How this addresses that need:

This removes the skip statements from the tests in question, so they
will run again.

Side effects of this change:

None.