Add instructions on how to confirm functionality
Why these changes are being introduced:

We've introduced a section in our readmes to confirm the app
is working after dependency updates, but we haven't yet done so
for TIMDEX UI.

Relevant ticket(s):

* [TIMX-269](https://mitlibraries.atlassian.net/browse/TIMX-269)

How this addresses that need:

This adds said instructions to the TIMDEX UI readme.

Side effects of this change:

The readme did not contain information on ADRs, so that has been
added, too, as is our practice with our repos.