Reorder geo bounding box fields
Why these changes are being introduced:

Longitude appears before latitude in the bounding box fieldset,
but it is more conventional to list latitude first. This also
matches the order of the geodistance fieldset.

Relevant ticket(s):

* [GDT-256](https://mitlibraries.atlassian.net/browse/GDT-256)

How this addresses that need:

This reorders the bounding box fieldset to display in the following
order (from top to bottom, then left to right):
* Minimum latitude
* Maximum latitude
* Minimum longitude
* Maximum longtidue

Side effects of this change:

In mobile, the fields will render in the order they are listed
in the markup; i.e., min lat, min long, max lat, max long. We
should confirm that this is acceptable.