Render filter value names in non-GDT contexts
Why these changes are being introduced:

When introducing the helper method to rename GDT source filter
values, we inadvertently removed all filter values for non-GDT
apps. This is hilarious, but not very usable.

Relevant ticket(s):

N/A

How this addresses that need:

This adds to the filter partial an `else` clause that renders the
term key if GDT is not enabled.

Side effects of this change:

Users will know what filters they are applying.