Render all possible fields in full record view
Why these changes are being introduced:

UXWS has requested that we dump all record metadata in a 'more
information' section, so we can gain insight in usability testing
as to which fields would be helpful to include int he full
record.

Relevant ticket(s):

* [GDT-253](https://mitlibraries.atlassian.net/browse/GDT-253)

How this addresses that need:

This refactors the record helper logic pertaining to the 'more
information' section, such that the fields that we attempt to
render in that section are only those that are mapped for geo
records in transmogrifier.

It also attempts to normalize how nested fields are rendered by
moving some of that logic to helper methods.

I still don't see this iteration as a permanent solution. My hope
is that we learn more in usability testing about which of these
fields are actually useful, so we can take more care in rendering
them properly and reduce that section of the full record. Darcy
has suggested that something rudimentary is acceptable at this point.

Side effects of this change:

* The `more_info` partial has been renamed to `more_info_geo`, to
clarify which full record view it should be applied to. This also
applies to the corresponding helper method.
* I've removed unused helper methods (and corresponding tests) that
I'm now fairly sure will no longer be used.
* I'm not sure that the markup in the `more_info_geo` partial is
semantically optimal. A list of lists raises some concerns for me,
and I wonder if each field name should be a heading instead.