Add publishers to geospatial full records
Why these changes are being introduced:

Now that the `publishers` field is available in the API via
recordId queries, we should add it to the metadata dump in GDT
full records.

Relevant ticket(s):

* [GDT-271](https://mitlibraries.atlassian.net/browse/GDT-271)
* [GDT-270](https://mitlibraries.atlassian.net/browse/GDT-270)

How this addresses that need:

This adds `publishers` to the `more_info_geo` partial and
corresponding helper test.

Side effects of this change:

None.