Display filters in search summary in order of application
Why these changes are being introduced:

The search summary currently displays filters in alphabetical
order by key. This is confusing, as one would expect them to
show up in the order they were applied.

Relevant ticket(s):

* [GDT-224](https://mitlibraries.atlassian.net/browse/GDT-224)

How this addresses that need:

Many (if not all) Rails helpers pertaining to URLs (`to_query`,
`to_params`, routing helpers, etc.) sort their inputs
lexicographically. This means that we cannot use any of these
methods, or the `params` object, to pass query params into
the Enhancer, as they will not remain ordered and will instead
be alphabetized everywhere in the application.

This commit seeks to fix this problem by passing
`request.original_url` into the Enhancer and building filter URLs
manually. It's messier than using the Rails convenience
methods, particularly because we're mpw mostly working with
strings rather than hashes, but it felt less complex than the
other approach considered of storing filter order in a cookie.

Prior to this change, both the Enhancer and QueryBuilder iterated
over arrays of supported params and added them to the query if
they showed up in the `params` object. This is now reversed: the
extraction methods in both models now iterate over the URL params
and confirm that they exist in the list of supported params before
adding them to the query. The reason for this change is to ensure
that filters (and other params) are added to the query in the order
in which they appear in the original URL, not the order in which
they are listed in the array of supported params.

Side effects of this change:

* Fixes a bug in the filter helper that prevents single-valued
filters from being applied. Currently, the only single-valued
filter we have is `literaryFormFilter`, which is not used anywhere,
but it makes sense to support all fields that TIMDEX API provides.
* Updates several tests to accommodate the refactored Enhancer
and QueryBuilder.
* Moves the constants that recur in the Enhancer and QueryBuilder
into a new module, QueryElements.
* Moves all of the filter helper methods related to adding/removing
filters next to each other for better readability.
* Regenerates several cassettes. (It's possible that not all
of these needed to be regenerated; I erred on the side of more.)