Improve page titles
Why these changes are being introduced:

Every view currently has the same generic title content. It's
common practice in discovery applications to include context about
the current query in the page title, so users can see these
pertinent details in a given tab or bookmark.

Relevant ticket(s):

* [GDT-234](https://mitlibraries.atlassian.net/browse/GDT-234)

How this addresses that need:

This adds helper methods to construct page titles for the following
views:

* `basic_search/index`: "Search `PLATFORM_NAME` | MIT Libraries"
* `record/view`: "`record['title']` | `PLATFORM_NAME` | MIT Libraries"
* `results`: "`params from enhanced query` | `PLATFORM_NAME` | MIT Libraries"

If the query/title preview string exceeds a configurable character
limit (defaults to 50 for results and 25 for record), the string
is truncated with an ellipsis.

Side effects of this change:

None.