Add affordance to clear search terms
Why these changes are being introduced:

Now that we have several input fields in the search form, it would
be helpful to have a button that can clear everything in the form.

Relevant ticket(s):

* [GDT-244](https://mitlibraries.atlassian.net/browse/GDT-244)

How this addresses that need:

This adds a 'clear' button to the form, with a corresponding
click handler that triggers a method to clear all inputs,
except hidden ones.

Side effects of this change:

This uses the `querySelectorAll()` method, which is
[supported by all modern browsers](https://developer.mozilla.org/en-US/docs/Web/API/Document/querySelectorAll#browser_compatibility)
but may not be supported by legacy browsers. This feels like an
acceptable risk.

An alternative approach is `getElementsByTagName`. This would also
select the hidden elements, which toggle geospatial and advanced
searches. I've avoided this because it may lead to some unexpected
behavior, but it also feels relatively low-risk.