Add turbo:load event listener to Matomo code
Why these changes are being introduced:

It’s a known issue that the out-of-the-box Matomo tracking code
does not work with Turbo, but we want to continue using Turbo
because it's great.

Relevant ticket(s):

* [GDT-289](https://mitlibraries.atlassian.net/browse/GDT-289)

How this addresses that need:

This adds a JS snippet to the tracking code, based on Matomo's
instructions for tracking SPA's, that sends an event to Matomo when
Turbo loads a page. See [this GitHub issue](https://github.com/turbolinks/turbolinks/issues/436)
for details.

Side effects of this change:

We should add this change to the theme gem if it works for apps
that are not using Turbo. (It should, but I haven't confirmed it.)