Add affordance to clear search terms
Why these changes are being introduced:

Now that we have several input fields in the search form, it
would be helpful to have a button that can clear everything in the
form.

Relevant ticket(s):

* [GDT-244](https://mitlibraries.atlassian.net/browse/GDT-244)

How this addresses that need:

This adds to the form an `a` tag, styled as a button, that links
to the root route.

Side effects of this change:

* In order to accommodate both buttons next to the basic search
input, the form breaks at a medium screen width rather than a small
one.
* Darcy has indicated that we may need to refine this as additional
use cases emerge.