Remove additional skip links
Why these changes are being introduced:

Kate and Rich from DAS advised us that skip links are primarily
for sighted users navigating with a keyboard, and it's very
unusual to find more than one skip link. To that end, we've
decided to simplify the skip links.

Relevant ticket(s):

* [GDT-280](https://mitlibraries.atlassian.net/browse/GDT-280)

How this addresses that need:

* All skip links now have the same conventional text: 'skip to main
content'.
* On the results view, there is one skip link that goes to the
results list.
* On the full record view, there is one skip link that goes to the
top of the record.
* On all other views (i.e., the landing page), there is one skip
link that goes to the first input in the search form (keyword
search).

Side effects of this change:

We may need to revisit this after a11y testing to confirm that
the skip links are useful.