Refactor markup to provide more semantic meaning
Why these changes are being introduced:

It was suggested in a recent meeting with DAS that the best
way to make a page navigable to screen reader users is by
using semantic HTML and ARIA roles.

Relevant ticket(s):

* [GDT-279](https://mitlibraries.atlassian.net/browse/GDT-279)

How this addresses that need:

This attempts to make better use of semantic HTML elements and ARIA
roles throughout our views.

One notable change is the addition of the `application` layout from
our theme. The layout has been added in a separate commit to clarify
what changes were made.

Customization of this layout is needed because the it adds a `main`
tag that surrounds all content between the nav and footer. While this
corresponds to the main content for some apps, it is less useful in
TIMDEX UI.

Side effects of this change:

* This should be considered a first iteration, as we have not
undergone a11y review yet. I tried to use a light touch due to my
lack of familiarity with ARIA, but we may learn that additional
context would be useful.
* We may decide in the future to make the changes to the `application`
layout upstream, in the theme gem. This would require making corresponding
changes in all of our Rails apps, so it is out of scope of this ticket.
* The full record sidebar is an `aside` in non-GDT apps, but in GDT it
includes the access link. To accommodate this, we've given it a
`role=region` with an `aria-label` mentioning the access link.
Long-term, however, we should consider moving this button outside of
the sidebar, so it has more prominence in the markup.