Restructure headings
Why these changes are being introduced:

We currently have two issues with our heading structure:

1. Multiple `h1` tags, which is not recommended.
2. Each result title is an `h2`, which is no longer semantically
correct as the results list now has an `h2` tag at the top.

Relevant ticket(s):

* [GDT-293](https://mitlibraries.atlassian.net/browse/GDT-293)

How this addresses that need:

This changes the `site_title` heading from an `h1` to an `h2`,
and updates results to use an `h3` heading for the record title,
and an `h4` heading for the search highlights section. It also
changes the 'Ask us' heading to an `h2`, so it doesn't become
nested beneath the filters heading.

Side effects of this change:

The `h1` in the header seems adequate for now, but we might want
to iterate on this further if DAS feels otherwise during
accessibility testing.