Display search terms in summary panel
Why these changes are being introduced:

We had been leaning on the search form to remind users of the
current search params, but this is not always reliable (e.g., if
a input is cleared without executing a new search).

Relevant ticket(s):

* [GDT-285](https://mitlibraries.atlassian.net/browse/GDT-285)

How this addresses that need:

This displays applied search terms in the summary panel where
we show applied filters.

Side effects of this change:

* Two skipped tests, which confirmed an earlier iteration of this
feature, have been reintroduced.
* The new param `booleanType` has been ignored in the page title
constructor as well as the summary panel.