Fix spacing for search summary panel
Why these changes are being introduced:

The current CSS rules for the search summary panel results in an
unintended top margin if geospatial/advanced search terms are
entered without a keyword.

Relevant ticket(s):

N/A

How this addresses that need:

This puts the top margin for applied terms on the `ul` elements
instead of the `li` elements. It also adjusts the margin to match
the others.

Side effects of this change:

N/A