Fact panels controlled via configuration
Why are these changes being introduced:

* Our experimental fact panels need to be disabled by default now that
  a version of this UI is moving to production

Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/GDT-313

How does this address that need:

* Adds a configuration option that takes a comma separated list of fact
  panels that should be enabled.
* If a fact panel is enabled and there is a a detected fact of that type
  a fact panel will display as it did previously. If a fact is detected
  of a type in which no fact panel has been enabled, we do not display
  a fact panel.
* If no fact panels are enabled, we don't check for specific fact panels
  to determine if they should be enabled.

Document any side effects to this change:

After discussion with the team, we chose not to use FlipFlop for this
work. We'll be considering over time whether the approach taken here is
a good option instead of reaching for FlipFlop at least in situations
where the feature does not extend into many places in the codebase.