Reorganize GeoData full record based on usability test results
Why these changes are being introduced:

Now that we've done usability testing, we have a better sense
of what fields in the 'More information' section of the full
record may be useful.

Relevant ticket(s):

* [GDT-283](https://mitlibraries.atlassian.net/browse/GDT-283)

How this addresses that need:

This moves each relevant field out of the 'more_info_geo' partial
and into the 'record_geo' partial. The display of each field has
been cleaned up a bit (e.g., removing subfield keys), and the
fields are now in `p` tags or `ul` tags with their own headings.

Since place names are duplicated as subjects, the locations field
has been repurposed to show only geoshapes. This also required
replacing the deprecated `geopoint` with `geoshape` in the
TimdexRecord query string, and regenerating the record controller
cassettes accordingly.

Some record helper logic (and corresponding tests) that are no
longer needed have also been removed.

Side effects of this change:

I'm not sure it's useful to display this metadata, particularly
when we provide an affordance to download the source metadata
and/or link to the source record. At some point, I'd like to have
a broader conversation about whether we should include full records
at all in TIMDEX UI. For now, I'm comfortable implementing this
as requested.