Improve accessibility of search form
Why these changes are being introduced:

Accessibility testing revealed that the search button is difficult
for screen reader users to locate, and that the `title` attribute for
the keyword search input is not useful.

Relevant ticket(s):

* [GDT-318](https://mitlibraries.atlassian.net/browse/GDT-318)

How this addresses that need:

This changes the keyword `title` to 'keyword anywhere' and moves
the search button to the bottom of the form, after all of the
advanced search dropdowns. It also updates some corresponding
CSS, including removing the `border-radius` override for the `btn`
class, so the search button looks more like a button.

Side effects of this change:

Fixed margin/padding for results summary panel and mobile filter
toggle. (This is unrelated to this change, but I noticed while
testing out the changes that the spacing looked odd.)