Consolidate full record access links in sidebar
Why these changes are being introduced:

The metadata download link appears at the end of the full record
with no heading. This makes it look like part of the last full
record heading (usually 'Languages').

A related issue is that the access link in the sidebar has no
heading. We have an `aria-label` that describes this section for
screen reader users, but for sighted users, this section lacks
context.

As part of their accessibility review, DAS has recommended moving
the metadata download link to the same section as the access link.

Relevant ticket(s):

* [GDT-319](https://mitlibraries.atlassian.net/browse/GDT-319)

How this addresses that need:

This moves the download link below the access link in the sidebar,
and adds a heading to that section.

Side effects of this change:

While working on this, I noticed that the fact panel markup needs
some work to be semantically meaningful. I haven't touched that in
this commit, but it will be worth revisiting when we return to
search intent detection.