Redesign authors partial for GDT
Why these changes are being introduced:

The authors partial lists all of the authors in a `p` tag, delimited
with a semicolon. An unordered list would be more semantically
meaningful.

UXWS has also suggested that we should include contributor types
in the list, to make clearer what their contribution to the work.

Relevant ticket(s):

* [GDT-284](https://mitlibraries.atlassian.net/browse/GDT-284)

How this addresses that need:

This refactors the authors partial to be an unordered list and
renders the contributor type alongside the name value.

Side effects of this change:

This does not touch the non-GDT record view, as it uses somewhat
different logic to render contributors. However, this might be
an opportunity to normalize, unless there's a valid reason why
the logic differs.