Address code review feedback
Matt noted two issues with the changeset:

1. It unintentionally removes the `truncated-list` class from the
contributors list in the results.
2. The `list-inline` class provided by the style library styles
list items with `inline-block`, not `inline`.

When I went to fix issue 1, I realized that issue 2 impacts it:
i.e., if the list items are `inline-block`, then `line-clamp` will
not truncate them properly.

This commit reintroduces `truncated-list` to the contributors list
in the results views and adds a rule to display items inline.
(The `list-inline` class is still applied, as it also removes the
left margin from the `ul`.)

This also changes how contributor kinds are rendered by refactoring
that logic to a helper method and handling some edge cases. This
is unrelated to the code review discussion, but I noticed it while
fixing the CSS issues.