Accommodate multi-digit list markers
Why these changes are being introduced:

The result list does not have a sufficient left margin to
accommodate list markers that are more than one digit. Because
of this, markers for results further down in the list are
truncated.

This issue exists on smaller viewports only. Larger viewports
have enough space for markers up to five digits. (This is the
maximum, in our case.)

Relevant ticket(s):

* [GDT-314](https://mitlibraries.atlassian.net/browse/GDT-314)

How this addresses that need:

This adds a left margin on smaller viewports to accommodate
4-digit result markers, and removes the result padding to provide
more room for text flow.

Side effects of this change:

* The margin is somewhat excessive for 1- or 2-digit numbers, but
it seems like a reasonable tradeoff for now.
* The leading digit of 5-digit numbers is still truncated. This only
affects one possible result, and it's unlikely that a user would get
to result 10,000.