Improve markup for contributors
Why these changes are being introduced:

Authors are currently rendered within a `p` tag. This is not very
semantically meaningful.

Relevant ticket(s):

* [GDT-284](https://mitlibraries.atlassian.net/browse/GDT-284)

How this addresses that need:

This refactors the contributors section to a partial and renders
contributors as an unordered list. It is the same changeset that
was introduced in #191, but without displaying the contributor type,
which UXWS has deemed unnecessary at this time.

Side effects of this change:

None.